+++ R/supports.R
@@ -22,7 +22,7 @@
 #' * `supports_pdftk()` detects support for the command-line tool `pdftk` which is
 #'   required for [get_bookmarks_pdftk()], [set_bookmarks_pdftk()],
 #'   [get_docinfo_pdftk()], [set_docinfo_pdftk()], [cat_pages_pdftk()], and [n_pages_pdftk()].
-#' * `requireNamespace("pdftools", quietly = TRUE)` detects support for the R packages `pdftools`
+#' * `requireNamespace("pdftools", quietly = TRUE)` detects support for the R package `pdftools` (>= 3.8.0)
 #'   which is required for [get_bookmarks_pdftools()] and [get_docinfo_pdftools()].
 #' * `requireNamespace("qpdf", quietly = TRUE)` detects support for the R packages `qpdf`
 #'   which is required for [cat_pages_qpdf()] and [n_pages_qpdf()].
@@ -153,6 +153,7 @@ supports_qpdf <- function() {
 }
 supports_pdftools <- function() {
 	requireNamespace("pdftools", quietly = TRUE) &&
+		utils::packageVersion("pdftools") >= "3.8.0" &&
 		!getOption("xmpdf_disable_pdftools", FALSE)
 }