+++ R/zzz.R
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: MIT
 
 #' @importFrom rlang abort inform warn %||%
-#' @importFrom utils hasName
+#' @importFrom utils hasName packageVersion
 #' @importFrom R6 R6Class
 #' @importFrom stringi stri_join
 NULL